refactor(wash-cli): reorganize `wash up` and `wash config`
This commit starts the reorganization of `wash-cli` code into the
`src/cmd` hierarchy, which should make it easier to separate shared
functionality, (sub)command invocation and specific logic.
