+++ crates/wash-cli/src/cmd/mod.rs
@@ -1,0 +1,4 @@
+//! Commands that are exposed by `wash`
+
+pub mod config;
+pub mod up;
