+++ crates/wash-cli/src/cmd/up.rs
@@ -35,11 +35,23 @@ use wasmcloud_control_interface::{Client as CtlClient, ClientBuilder as CtlClien
 
 use crate::app::deploy_model_from_manifest;
 use crate::appearance::spinner::Spinner;
-use crate::down::stop_nats;
+use crate::config::{
+    configure_host_env, DEFAULT_ALLOW_FILE_LOAD, DEFAULT_LATTICE, DEFAULT_MAX_EXECUTION_TIME_MS,
+    DEFAULT_NATS_HOST, DEFAULT_NATS_PORT, DEFAULT_NATS_WEBSOCKET_PORT,
+    DEFAULT_PROV_SHUTDOWN_DELAY_MS, DEFAULT_RPC_TIMEOUT_MS, DEFAULT_STRUCTURED_LOG_LEVEL,
+    NATS_SERVER_VERSION, WADM_VERSION, WASMCLOUD_ALLOW_FILE_LOAD, WASMCLOUD_CLUSTER_ISSUERS,
+    WASMCLOUD_CLUSTER_SEED, WASMCLOUD_CONFIG_SERVICE, WASMCLOUD_CTL_CREDSFILE, WASMCLOUD_CTL_HOST,
+    WASMCLOUD_CTL_JWT, WASMCLOUD_CTL_PORT, WASMCLOUD_CTL_SEED, WASMCLOUD_CTL_TLS,
+    WASMCLOUD_CTL_TLS_CA_FILE, WASMCLOUD_ENABLE_IPV6, WASMCLOUD_HOST_LOG_PATH, WASMCLOUD_HOST_SEED,
+    WASMCLOUD_HOST_VERSION, WASMCLOUD_JS_DOMAIN, WASMCLOUD_LATTICE, WASMCLOUD_LOG_LEVEL,
+    WASMCLOUD_MAX_EXECUTION_TIME_MS, WASMCLOUD_OCI_ALLOWED_INSECURE, WASMCLOUD_OCI_ALLOW_LATEST,
+    WASMCLOUD_POLICY_TOPIC, WASMCLOUD_PROV_SHUTDOWN_DELAY_MS, WASMCLOUD_RPC_CREDSFILE,
+    WASMCLOUD_RPC_HOST, WASMCLOUD_RPC_JWT, WASMCLOUD_RPC_PORT, WASMCLOUD_RPC_SEED,
+    WASMCLOUD_RPC_TIMEOUT_MS, WASMCLOUD_RPC_TLS, WASMCLOUD_RPC_TLS_CA_FILE,
+    WASMCLOUD_SECRETS_TOPIC, WASMCLOUD_STRUCTURED_LOGGING_ENABLED,
+};
 
-mod config;
-mod credsfile;
-pub use config::*;
+use crate::down::stop_nats;
 
 #[derive(Parser, Debug, Clone)]
 pub struct UpCommand {
