+++ crates/wash-cli/src/lib.rs
@@ -1,7 +1,13 @@
+//! Functionality that powers the `wash` CLI
+
+pub(crate) mod creds;
+pub(crate) mod errors;
+
 pub mod app;
 pub mod appearance;
 pub mod build;
 pub mod call;
+pub mod cmd;
 pub mod common;
 pub mod completions;
 pub mod config;
@@ -17,5 +23,4 @@ pub mod plugin;
 pub mod secrets;
 pub mod style;
 pub mod ui;
-pub mod up;
 pub mod util;
