+++ crates/wash-cli/tests/wash_config.rs
@@ -4,7 +4,7 @@ use common::TestWashInstance;
 
 use std::collections::HashMap;
 
-use wash_cli::config::ConfigCliCommand;
+use wash_cli::cmd::config::ConfigCliCommand;
 use wash_lib::cli::{CliConnectionOpts, OutputKind};
 
 #[tokio::test]
@@ -23,21 +23,22 @@ async fn test_config_put_and_get() -> anyhow::Result<()> {
     };
 
     // Put the config
-    wash_cli::config::handle_command(command, OutputKind::Json).await?;
+    wash_cli::cmd::config::handle_command(command, OutputKind::Json).await?;
 
     // Assert that the retrieved config deserializes as a HashMap
-    let retrieved_config: HashMap<String, serde_json::Value> = wash_cli::config::handle_command(
-        ConfigCliCommand::GetCommand {
-            opts: CliConnectionOpts {
-                ctl_port: Some(wash_instance.nats_port.to_string()),
-                ..Default::default()
+    let retrieved_config: HashMap<String, serde_json::Value> =
+        wash_cli::cmd::config::handle_command(
+            ConfigCliCommand::GetCommand {
+                opts: CliConnectionOpts {
+                    ctl_port: Some(wash_instance.nats_port.to_string()),
+                    ..Default::default()
+                },
+                name: "foobar".to_string(),
             },
-            name: "foobar".to_string(),
-        },
-        OutputKind::Json,
-    )
-    .await?
-    .map;
+            OutputKind::Json,
+        )
+        .await?
+        .map;
 
     assert_eq!(retrieved_config.len(), 2);
     assert_eq!(retrieved_config.get("key").unwrap(), "value");
@@ -58,7 +59,7 @@ async fn test_config_secret_name_error() -> anyhow::Result<()> {
     };
 
     // Put the config and expect an error
-    let result = wash_cli::config::handle_command(command, OutputKind::Json).await;
+    let result = wash_cli::cmd::config::handle_command(command, OutputKind::Json).await;
 
     // Assert that an error occurred
     assert!(result.is_err());
